poetry -> uv, upgrade dependencies, toml config, turn into package
