--- config_example.py
@@ -1,7 +1,0 @@
-PREFIX = "!"
-TOKEN = ""
-# Todo -- update class dictionary
-BLACKLIST_COGS = {}
-
-NO_DOCKER = False
-USE_ONLINETEX = False
