+++ discordmanimator/cogs/search_doc.py
